Moved stuff in different folders (WIP)
