+++ src/modules/WindowsSpecsCheck.py
