+++ src/modules/networkAdapter.py
@@ -1,0 +1,4 @@
+class NetworkAdapter:
+    def __init__(self, enabled: bool, type: str) -> None:
+        self.enabled = enabled
+        self.type = type
